Framebuffer bug fixes
When you are drawing a Framebuffer, the screen was not clear between each render call so the screen was filling with with colors of surfaces and things like that.
The Framebuffer is still work in progress because the default size is not correct.
The Shader was creating a new FloatBuffer each time an uploadMat3f was called, it is notg the case anymore, it clear the buffer and reuse it (this was causing memory leakage).