Slider implementation and render groups
I have reviewed the GameObject definition and we were using GameObjects as RenderGroups which is not correct so I renamed the GameObject object to RenderGroup and I will implement a real GameObject when I am going to do Collider and RigidBodies.
The Slider is now working graphically, now I have to implement the numbers things.