Window name
Renamed the window name from Don't Play Together V2 to DSEngine because it is more appropriate. In fact the window name is now stored into a String variable that you can get with the getName(). Before the name of the window was hard coded which is really a dumb idea =')
Removed the 'disney' background because it can hide some unwanted features.